requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6; sys_platform == 'win32'
comtypes==1.4.10; sys_platform == 'win32'
contourpy==1.3.1
cycler==0.12.1
numpy==2.2.4
Flask==3.1.0
Flask-Caching==2.3.1
flask-cors==5.0.1
fonttools==4.56.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
MouseInfo==0.1.3; sys_platform == 'win32'
multidict==6.2.0
narwhals==1.31.0
nest-asyncio==1.6.0
nltk==3.9.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
pyparsing==3.2.1
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.1
pywhatkit==5.4
rapidfuzz==3.12.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.29.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.14.1
streamlit==1.43.2
tenacity==9.0.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==3.1.3
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.18.3
gunicorn
